Update __init__.py
I had problems installing the main branch using Ubuntu 18.04 LTS and Anaconda. More precisely, following the Mac instructions, astrochem installed, but would raise an error when first imported. I *think* there was a change in the default behavior of looking for modules in python packages somewhere around python 2.7. Previously, relative paths were okay. More recently, absolute paths are expected and/or required. That is background. The symptom was that 
> import astrochem
resulted in an error. __init__.py ran, but could not find tools, wrapper or libpyastrochem. The solution was to change
import tools
to
from . import tools
and the equivalent for wrapper and libpyastrochem